+++ app/package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "@kibalabs/core": "0.5.7-next.1",
-    "@kibalabs/core-react": "^0.7.1-next.0",
+    "@kibalabs/core-react": "^0.7.1-next.2",
     "@kibalabs/everyview-tracker": "^0.2.5",
     "@kibalabs/everyview-tracker-react": "^0.2.5",
     "@kibalabs/ui-react": "^0.9.6-next.8",
